Fix DB schema mismatch
Schema-validation: wrong column type encountered in column [start_day] in table [tour]; found [smallint (Types#SMALLINT)], but expecting [integer (Types#INTEGER)]

+++ src/main/kotlin/de/cyclingsir/cetrack/tour/domain/DomainMTTour.kt
@@ -28,11 +28,11 @@ data class DomainMTTour(
 
     val STARTTIMESTAMP: Long,
 
-    val STARTYEAR: Int,
+    val STARTYEAR: Short,
 
-    val STARTMONTH: Int,
+    val STARTMONTH: Short,
 
-    val STARTDAY: Int,
+    val STARTDAY: Short,
 
     val TOURALTUP: Int,
